refactor: [EL-4914] reorganize interactive tours module structure and update imports
