+++ src/[fsd]/app/layout/AppLayout.jsx
@@ -5,7 +5,7 @@ import { Box } from '@mui/material';
 import MainPanel from '@/[fsd]/app/layout/MainPanel';
 import MainSidebar from '@/[fsd]/app/layout/MainSidebar';
 import { InteractiveTourProvider } from '@/[fsd]/app/providers/InteractiveTourProvider';
-import { useTourFromUrl } from '@/[fsd]/features/interactive-tours/model/useTourFromUrl';
+import { useTourFromUrl } from '@/[fsd]/features/interactive-tours';
 import { SupportAssistantWidget } from '@/[fsd]/widgets/SupportAssistant';
 
 const AppLayoutInner = memo(props => {
