+++ src/[fsd]/app/providers/InteractiveTourContext.js
@@ -1,3 +1,5 @@
-import { createContext } from 'react';
+import { createContext, useContext } from 'react';
 
 export const InteractiveTourContext = createContext(null);
+
+export const useInteractiveTour = () => useContext(InteractiveTourContext);
