+++ src/[fsd]/features/interactive-tours/lib/constants/index.js
@@ -1,0 +1,1 @@
+export { CARD_WIDTH_PX } from './interactiveTour.constants';
