+++ src/[fsd]/features/interactive-tours/lib/constants/interactiveTour.constants.js
@@ -1,0 +1,1 @@
+export const CARD_WIDTH_PX = 440;
