+++ src/[fsd]/features/interactive-tours/lib/hooks/index.js
@@ -1,0 +1,3 @@
+export { isTourCompleted, useProposeTour } from './useProposeTour.hooks';
+export { useTourCardPosition } from './useTourCardPosition.hooks';
+export { useTourFromUrl } from './useTourFromUrl.hooks';
