+++ src/[fsd]/features/interactive-tours/lib/hooks/useTourFromUrl.hooks.js
@@ -2,7 +2,7 @@ import { useEffect, useRef } from 'react';
 
 import { useSearchParams } from 'react-router-dom';
 
-import { useInteractiveTour } from './useInteractiveTour';
+import { useInteractiveTour } from '@/[fsd]/app/providers/InteractiveTourContext';
 
 export const useTourFromUrl = () => {
   const [searchParams, setSearchParams] = useSearchParams();
