--- src/[fsd]/features/interactive-tours/model/useInteractiveTour.js
@@ -1,5 +1,0 @@
-import { useContext } from 'react';
-
-import { InteractiveTourContext } from '@/[fsd]/app/providers/InteractiveTourContext';
-
-export const useInteractiveTour = () => useContext(InteractiveTourContext);
